refactor(connectors): move connector alert helpers into onyx/db
db modules were importing onyx/notifications, inverting the layering;
the helpers are thin wrappers over db functions and belong in onyx/db.
Also document that the alert producer rolls back on failure.
